fix: fix `update-pre-commit-hooks` workflow
The `update-pre-commit-hooks` workflow used `REPO_PAT` as its token,
which caused it fail with the default permissions suggested for the
token (repo restricted 'contents:write' and 'pull_requests:write').
Custom tokens, used to allow pull requests to trigger workflows, need to
be global. This commit simply removes the dependence on `REPO_PAT`, and
uses the default `GITHUB_TOKEN` in the workflow--consequently the pull
request made by `update-pre-commit-hooks` will not trigger `check-pr`.
Additionally, 'pull_requests:write' is no longer listed as a required
permission for `REPO_PAT`.